Add first-class adoption CLI command and wire follow-up workflow
