+++ docs/decision-guide.md
@@ -28,6 +28,7 @@ Convert fit + latest gate decision into prioritized follow-up actions:
 ```bash
 make adoption-followup
 make adoption-followup-contract
+python -m sdetkit adoption --format json
 ```
 
 ## 2) When SDETKit is probably *not* worth it
