+++ src/sdetkit/parsed_shortcuts.py
@@ -20,6 +20,7 @@ _PARSED_MODULE_SHORTCUTS: tuple[tuple[str, str], ...] = (
     ("kv", "sdetkit.kvcli"),
     ("evidence", "sdetkit.evidence"),
     ("onboarding", "sdetkit.onboarding"),
+    ("adoption", "sdetkit.adoption"),
     ("start", "sdetkit.onboarding"),
     ("onboarding-optimization", "sdetkit.onboarding_optimization"),
     ("community-activation", "sdetkit.community_activation"),
